pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "innie-engine"
version = "0.1.0"
description = "Persistent memory and identity for AI coding assistants"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
authors = [{ name = "Joshua Jenquist" }]
keywords = ["ai", "memory", "cli", "claude", "coding-assistant", "knowledge-base"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Application Frameworks",
]

dependencies = [
    "typer>=0.15",
    "rich>=13.0",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "jinja2>=3.1",
    "httpx>=0.27",
    "sqlite-vec>=0.1",
]

[project.optional-dependencies]
serve = [
    "fastapi>=0.115",
    "uvicorn>=0.32",
    "aiosqlite>=0.21",
]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.8",
    "fastapi>=0.115",
    "uvicorn>=0.32",
    "aiosqlite>=0.21",
]
docs = [
    "mkdocs>=1.6",
    "mkdocs-material>=9.5",
    "pymdown-extensions>=10.0",
]

[project.scripts]
innie = "innie.cli:app"

[project.entry-points."innie.backends"]
claude-code = "innie.backends.claude_code:ClaudeCodeBackend"
opencode = "innie.backends.opencode:OpenCodeBackend"
cursor = "innie.backends.cursor:CursorBackend"

[project.urls]
Homepage = "https://github.com/joshuajenquist/innie-engine"
Repository = "https://github.com/joshuajenquist/innie-engine"
Issues = "https://github.com/joshuajenquist/innie-engine/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/innie"]

[tool.hatch.build.targets.wheel.shared-data]
"src/innie/hooks" = "innie/hooks"
"src/innie/templates" = "innie/templates"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
